requirements.txt
uvicorn[standard]
fastapi
pydantic[Email]
python-dotenv==1.0.1
pymongo==4.7.2
passlib==1.7.4
bcrypt
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
motor==3.4.0
pyjwt==2.8.0
redis==5.0.4
langchain==0.2.15
langchain-ollama==0.1.3
ollama==0.3.2
langchain-community==0.2.15
pypdf==4.3.1
langchain-chroma==0.1.1
langchain-text-splitters==0.2.2
chromadb==0.5.5
boto3
nltk==3.9.1
pandas
scikit-learn==1.5.1
textblob
beanie